Refactored FlutterSecureStorage username/password storage into global functions
- Used to be located inside of UserDataProvider
- This is to assist with the effort to migrate AuthenticationModel to a query hook
