+++ lib/core/hooks/parking_query.dart
@@ -1,9 +1,11 @@
 import 'package:campus_mobile_experimental/core/models/parking.dart';
 import 'package:fquery/fquery.dart';
 
-UseQueryResult<List<ParkingModel>, dynamic> useFetchParkingModels()
+// STAGE 1 - LOADING AND INIT
+UseQueryResult<List<ParkingModel>, dynamic> useFetchParkingModels(String pid)
 {
   return useQuery(['parking'], () async {
+    print (pid);
     throw UnimplementedError("PARKING MODEL FQUERY HOOK NOT IMPLEMENTED");
   });
 }
