+++ lib/core/providers/user.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'dart:typed_data';
 
 import 'package:campus_mobile_experimental/core/models/authentication.dart';
 import 'package:campus_mobile_experimental/core/models/user_profile.dart';
@@ -7,14 +6,10 @@ import 'package:campus_mobile_experimental/core/providers/cards.dart';
 import 'package:campus_mobile_experimental/core/providers/notifications.dart';
 import 'package:campus_mobile_experimental/core/services/authentication.dart';
 import 'package:campus_mobile_experimental/core/services/user.dart';
-import 'package:encrypt/encrypt.dart';
+import 'package:campus_mobile_experimental/core/utils/user_credentials.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:hive/hive.dart';
-import 'package:pointycastle/asymmetric/api.dart';
-import 'package:pointycastle/asymmetric/oaep.dart';
-import 'package:pointycastle/pointycastle.dart' as pc;
 
 class UserDataProvider extends ChangeNotifier {
   UserDataProvider() {
@@ -24,7 +19,6 @@ class UserDataProvider extends ChangeNotifier {
     ///INITIALIZE SERVICES
     _authenticationService = AuthenticationService();
     _userProfileService = UserProfileService();
-    storage = FlutterSecureStorage();
 
     ///default authentication model and profile is needed in this class
     _authenticationModel = AuthenticationModel.fromJson({});
@@ -40,7 +34,6 @@ class UserDataProvider extends ChangeNotifier {
   ///MODELS
   AuthenticationModel? _authenticationModel;
   UserProfileModel? _userProfileModel;
-  late FlutterSecureStorage storage;
 
   ///SERVICES
   late AuthenticationService _authenticationService;
@@ -123,58 +116,6 @@ class UserDataProvider extends ChangeNotifier {
     notifyListeners();
   }
 
-  /// Save encrypted password to device
-  void _saveEncryptedPasswordToDevice(String encryptedPassword) {
-    storage.write(key: 'encrypted_password', value: encryptedPassword);
-  }
-
-  /// Get encrypted password that has been saved to device
-  Future<String?> _getEncryptedPasswordFromDevice() {
-    return storage.read(key: 'encrypted_password');
-  }
-
-  /// Save username to device
-  void _saveUsernameToDevice(String username) {
-    storage.write(key: 'username', value: username);
-  }
-
-  /// Get username from device
-  Future<String?> getUsernameFromDevice() {
-    return storage.read(key: 'username');
-  }
-
-  /// Delete username from device
-  void _deleteUsernameFromDevice() {
-    storage.delete(key: 'username');
-  }
-
-  /// Delete password from device
-  void _deletePasswordFromDevice() {
-    storage.delete(key: 'password');
-  }
-
-  /// Encrypt given username and password and store on device
-  void _encryptAndSaveCredentials(String username, String password) {
-    // TODO: import assets/public_key.txt
-    final String pkString = '-----BEGIN PUBLIC KEY-----\n' +
-        'MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDJD70ejMwsmes6ckmxkNFgKley\n' +
-        'gfN/OmwwPSZcpB/f5IdTUy2gzPxZ/iugsToE+yQ+ob4evmFWhtRjNUXY+lkKUXdi\n' +
-        'hqGFS5sSnu19JYhIxeYj3tGyf0Ms+I0lu/MdRLuTMdBRbCkD3kTJmTqACq+MzQ9G\n' +
-        'CaCUGqS6FN1nNKARGwIDAQAB\n' +
-        '-----END PUBLIC KEY-----';
-
-    final rsaParser = RSAKeyParser();
-    final pc.RSAPublicKey publicKey = rsaParser.parse(pkString) as RSAPublicKey;
-    var cipher = OAEPEncoding(pc.AsymmetricBlockCipher('RSA'));
-    pc.AsymmetricKeyParameter<pc.RSAPublicKey> keyParametersPublic =
-        new pc.PublicKeyParameter(publicKey);
-    cipher.init(true, keyParametersPublic);
-    Uint8List output = cipher.process(utf8.encode(password) as Uint8List);
-    var base64EncodedText = base64.encode(output);
-    _saveUsernameToDevice(username);
-    _saveEncryptedPasswordToDevice(base64EncodedText);
-  }
-
   /// Authenticate a user given an username and password
   /// Upon logging in we should make sure that users has an account
   /// If the user doesn't have an account one will be made by invoking [_createNewUser]
@@ -184,7 +125,7 @@ class UserDataProvider extends ChangeNotifier {
     notifyListeners();
 
     if (username.isNotEmpty && password.isNotEmpty) {
-      _encryptAndSaveCredentials(username, password);
+      encryptAndSaveCredentials(username, password);
 
       if (await silentLogin()) {
         if (_userProfileModel!.classifications!.student!) {
@@ -216,7 +157,7 @@ class UserDataProvider extends ChangeNotifier {
     notifyListeners();
 
     String? username = await getUsernameFromDevice();
-    String? encryptedPassword = await _getEncryptedPasswordFromDevice();
+    String? encryptedPassword = await getEncryptedPasswordFromDevice();
 
     /// Allow silentLogin if username, pw are set, and the user is not logged in
     if (username != null && encryptedPassword != null) {
@@ -258,8 +199,7 @@ class UserDataProvider extends ChangeNotifier {
         .unregisterDevice(_authenticationModel!.accessToken);
     updateAuthenticationModel(AuthenticationModel.fromJson({}));
     updateUserProfileModel(await _createNewUser(UserProfileModel.fromJson({})));
-    _deletePasswordFromDevice();
-    _deleteUsernameFromDevice();
+    deleteUserCredentialsFromDevice();
     CardsDataProvider _cardsDataProvider = CardsDataProvider();
     _cardsDataProvider.updateAvailableCards("");
     var box = await Hive.openBox<AuthenticationModel?>('AuthenticationModel');
